+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.1] - 2023-01-11
 
-## Added
+### Fixed
 
-- go.mod creation when running `lip tooth init`.
+- Failing to fetch tooth whose version has suffix `+incompatible`.
+- Failing to parse wildcards.
 
 ## [0.2.0] - 2023-01-11
 
@@ -35,7 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.2.1...HEAD
 [0.2.1]: https://github.com/LiteLDev/Lip/releases/tag/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.1.0